splitted LEDController into protocol and FastLED part
